+++ CHANGELOG.md
@@ -2,6 +2,8 @@ CHANGELOG for Sulu
 ==================
 
 * dev-develop
+    * BUGFIX      #1646 [ContactBundle]   Fixed upload of contact-avatar when a position is applied to the contact
+    * BUGFIX      #1645 [CategoryBundle]  Removed automatic category translation
     * BUGFIX      #1642 [AdminBundle]     Fixed cropping issue in table-view of datagrid
     * BUGFIX      #1641 [MediaBundle]     Fixed types in media-selection
     * ENHANCEMENT #1638 [ContentBundle]   Removed unnecessary variable in content form